Move thread to `src/gui`.
